convert temperature_celcius_to_fahrenheit() function to float
This allows for greater precision avoiding trucation errors when using int.

Tested with the following commands (before):

$ ./build/celtofah 1
33

$ ./build/celtofah 2
35

Tested with the following commands (after):

$ ./build/celtofah 1
33.8

$ ./build/celtofah 2
35.6

Also used DuckDuckGo converter tool [1] to verify these results.

[1] - https://duckduckgo.com/?t=ffab&q=celsius+to+fahrenheit&ia=web

+++ c/apps/celtofah.c
@@ -6,8 +6,8 @@
 int
 main(int argc, char **argv)
 {
-  int celcius = 0;
-  int fahrenheit = 0;
+  float celcius = 0;
+  float fahrenheit = 0;
 
   if (argc < 2)
     {
@@ -15,10 +15,10 @@ main(int argc, char **argv)
       return 1;
     }
 
-  celcius = atoi(argv[1]);
+  celcius = atof(argv[1]);
 
   fahrenheit = temperature_celcius_to_fahrenheit(celcius);
-  printf("%d\n", fahrenheit);
+  printf("%.1f\n", fahrenheit);
 
   return 0;
 }
