+++ c/apps/temperature.h
@@ -1,7 +1,7 @@
 #ifndef TEMPERATURE_H_INCLUDED_
 #define TEMPERATURE_H_INCLUDED_ 1
 
-int temperature_celcius_to_fahrenheit(int celcius);
+float temperature_celcius_to_fahrenheit(float celcius);
 float temperature_fahrenheit_to_celcius(float fahrenheit);
 
 #endif  /* !defined(TEMPERATURE_H_INCLUDED_) */
